Update README.md
We are creating a recommender system to identify ideal products for wholesale retailers using a historical dataset and purchase patterns of established customers to determine if we can recommend unique items that would be likely for retailers to purchase. This recommender system was created to see the efficiency for a wholesaler retailer provider and its impact on how items would be recommended and sold from items that may not be sold as frequently to encourage frequent inventory turnover.